+++ src/index.ts
@@ -10,6 +10,11 @@ export default JSXSlack
 export type Node = JSXSlack.ChildElements
 export type FunctionComponent<P extends {} = {}> = JSXSlack.FunctionComponent<P>
 export type FC<P extends {} = {}> = JSXSlack.FC<P>
-export type VoidFunctionComponent<P extends {} = Record<any, never>> =
+export type PropsWithChildren<P extends {} = {}> = JSXSlack.PropsWithChildren<P>
+
+/** @deprecated Use FunctionComponent instead. */
+export type VoidFunctionComponent<P extends {} = {}> =
   JSXSlack.VoidFunctionComponent<P>
-export type VFC<P extends {} = Record<any, never>> = JSXSlack.VFC<P>
+
+/** @deprecated Use FunctionComponent instead. */
+export type VFC<P extends {} = {}> = JSXSlack.VFC<P>