+++ test/jsx.tsx
@@ -1,10 +1,5 @@
 /** @jsx JSXSlack.h */
-import type {
-  FC,
-  FunctionComponent,
-  VFC,
-  VoidFunctionComponent,
-} from '../src/index'
+import type { FC, FunctionComponent } from '../src/index'
 import { JSXSlack } from '../src/jsx'
 import {
   createComponent,
@@ -263,9 +258,8 @@ describe('JSX', () => {
   })
 
   describe('Types', () => {
-    describe('JSXSlack.FunctionalComponent / (JSXSlack.)FunctionComponent / (JSXSlack.)FC', () => {
-      it('accepts specified props and children', () => {
-        const functionalComponent: JSXSlack.FunctionalComponent = () => null
+    describe('JSXSlack.FunctionComponent / JSXSlack.FC', () => {
+      it('accepts only specified props', () => {
         const functionComponent: JSXSlack.FunctionComponent = () => null
         const fc: JSXSlack.FC<{ test: string }> = ({ test }) => (
           <JSXSlack.Fragment>{test}</JSXSlack.Fragment>
@@ -273,54 +267,25 @@ describe('JSX', () => {
         const publicFunctionComponent: FunctionComponent<{ test: string }> = ({
           test,
         }) => <JSXSlack.Fragment>{test}</JSXSlack.Fragment>
-        const publicFC: FC<{}> = () => null // eslint-disable-line @typescript-eslint/ban-types
+        const publicFC: FC<Record<string, never>> = () => null
 
-        expect(functionalComponent({ children: [] })).toBeNull()
+        // // @ts-expect-error children prop is not allowed in FunctionComponent
         expect(functionComponent({ children: [] })).toBeNull()
+        expect(functionComponent({})).toBeNull()
+
+        // @ts-expect-error children prop is not allowed in FunctionComponent
         expect(fc({ test: 'abc', children: [] })).toStrictEqual(['abc'])
+        expect(fc({ test: 'abc' })).toStrictEqual(['abc'])
+
         expect(
+          // @ts-expect-error children prop is not allowed in FunctionComponent
           publicFunctionComponent({ test: 'def', children: [] })
         ).toStrictEqual(['def'])
+        expect(publicFunctionComponent({ test: 'def' })).toStrictEqual(['def'])
+
+        // @ts-expect-error children prop is not allowed in FunctionComponent
         expect(publicFC({ children: [] })).toBeNull()
-      })
-    })
-
-    describe('JSXSlack.VoidFunctionalComponent / (JSXSlack.)VoidFunctionComponent / (JSXSlack.)VFC', () => {
-      it('accepts only specified props', () => {
-        const voidFunctionalComponent: JSXSlack.VoidFunctionalComponent = () =>
-          null
-        const voidFunctionComponent: JSXSlack.VoidFunctionComponent = () => null
-        const vfc: JSXSlack.VFC<{ test: string }> = ({ test }) => (
-          <JSXSlack.Fragment>{test}</JSXSlack.Fragment>
-        )
-        const publicVoidFunctionComponent: VoidFunctionComponent<{
-          test: string
-        }> = ({ test }) => <JSXSlack.Fragment>{test}</JSXSlack.Fragment>
-        const publicVFC: VFC<Record<string, never>> = () => null
-
-        // @ts-expect-error children prop is not allowed in VoidFunctionalComponent
-        expect(voidFunctionalComponent({ children: [] })).toBeNull()
-        expect(voidFunctionalComponent({})).toBeNull()
-
-        // @ts-expect-error children prop is not allowed in VoidFunctionalComponent
-        expect(voidFunctionComponent({ children: [] })).toBeNull()
-        expect(voidFunctionComponent({})).toBeNull()
-
-        // @ts-expect-error children prop is not allowed in VoidFunctionalComponent
-        expect(vfc({ test: 'abc', children: [] })).toStrictEqual(['abc'])
-        expect(vfc({ test: 'def' })).toStrictEqual(['def'])
-
-        expect(
-          // @ts-expect-error children prop is not allowed in VoidFunctionalComponent
-          publicVoidFunctionComponent({ test: 'abc', children: [] })
-        ).toStrictEqual(['abc'])
-        expect(publicVoidFunctionComponent({ test: 'def' })).toStrictEqual([
-          'def',
-        ])
-
-        // @ts-expect-error children prop is not allowed in VoidFunctionalComponent
-        expect(publicVFC({ children: [] })).toBeNull()
-        expect(publicVFC({})).toBeNull()
+        expect(publicFC({})).toBeNull()
       })
     })
   })